Presenting the first version of LLM-code-porting project
This project aims to provide a small application
that implements LLM models in order to port (or
translate) Python code to C++, Rust and Java code
in order to improve execution time performance.
In this project I test different closed and open
LLM models in order to also assess which are the
best in porting the py code correctly.